Split filter name - change files names
