+++ src/ds5/ds5-device.cpp
@@ -36,7 +36,7 @@
 #include "proc/depth-formats-converter.h"
 #include "proc/depth-decompress.h"
 #include "proc/hdr-merge.h"
-#include "proc/filter-by-sequence-id.h"
+#include "proc/sequence_id_filter.h"
 #include "hdr-config.h"
 #include "../common/fw/firmware-version.h"
 #include "fw-update/fw-update-unsigned.h"
