+++ src/proc/CMakeLists.txt
@@ -21,7 +21,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/spatial-filter.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/temporal-filter.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/hdr-merge.cpp"
-        "${CMAKE_CURRENT_LIST_DIR}/filter-by-sequence-id.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/sequence_id_filter.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/hole-filling-filter.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/disparity-transform.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/y8i-to-y8y8.cpp"
@@ -48,7 +48,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/spatial-filter.h"
         "${CMAKE_CURRENT_LIST_DIR}/temporal-filter.h"
         "${CMAKE_CURRENT_LIST_DIR}/hdr-merge.h"
-        "${CMAKE_CURRENT_LIST_DIR}/filter-by-sequence-id.h"
+        "${CMAKE_CURRENT_LIST_DIR}/sequence_id_filter.h"
         "${CMAKE_CURRENT_LIST_DIR}/hole-filling-filter.h"
         "${CMAKE_CURRENT_LIST_DIR}/syncer-processing-block.h"
         "${CMAKE_CURRENT_LIST_DIR}/disparity-transform.h"
