+++ src/proc/sequence_id_filter.cpp
@@ -1,7 +1,7 @@
 // License: Apache 2.0. See LICENSE file in root directory.
 // Copyright(c) 2020 Intel Corporation. All Rights Reserved.
 
-#include "filter-by-sequence.h"
+#include "sequence_id_filter.h"
 
 namespace librealsense
 {
